Restored Map click modal functionality
